(naively) overriding namer and typer
On the day one of paradise 2.0, I figured that one can hijack analyzer
using reflection and then override its method to deeply customize
typechecking.

So when it didn't work today, I was scared to the bones. However the
solution was simple - one has to hijack the analyzer field on plugin
instantiation, not during the designated phase, and then it seems to work.

upd. Haha, it was too early to declare victory, because just changing the
analyzer field is not enough, because by the time the plugin is
instantiated, the old analyzer has already sprouted namerFactory,
packageObjects and typerFactory. Now THAT was scary.