work around weird test flakiness
It would seem that the repl in the “ad-hoc macros expand” test behaves differently
depending on the environment it’s been run in:
https://github.com/scala/community-builds/pull/529#issuecomment-294302008

Previously, the corresponding assert was checking that the resulting printout
is exactly equal to the expected output. Now, I changed it to only check
for the presence of key fragments of the expected output.